1.3.0
- Cambiata icona del programma
- Aggiunto il campo Visto (bool) per segnalare che lo abbiamo visto
- Al salvataggio del file di db cancella il db.bak in caso di salvataggio effettuato con successo
- Nella gestione files, l'eliminazione è diventata esclusione dall'archivio, e chiede se deve anche eliminare il file fisico
- Aggiunto il flag Visto ai TreeView della maschera di gestione dei files e di scelta dei files
- Corretto la funzione di scrittura del db in formato json per evitare errori nella creazione del file di backup se non esiste quello di db

+++ LettoreVideo/Classi/JsonOperation.cs
@@ -18,7 +18,8 @@ namespace LettoreVideo.Classi
 
                 if (Directory.Exists(pDataFolder))
                 {
-                    System.IO.File.Copy(DbFile, Path.Combine(pDataFolder, "db.bak"), true);
+                    if (File.Exists(DbFile))
+                        System.IO.File.Copy(DbFile, Path.Combine(pDataFolder, "db.bak"), true);
                 }
 
                 //string json = JsonConvert.SerializeObject(VID_DBs, Formatting.Indented);
@@ -36,9 +37,10 @@ namespace LettoreVideo.Classi
                     System.IO.File.Copy(Path.Combine(pDataFolder, "db.bak"), DbFile, true);
                 }
                 PRG.MsgBoxERR(ex, "Errore nella procedura di aggiornamento dell'archivio:\r\n\r\n");
+                
             }
 
-            if (Directory.Exists(Path.Combine(pDataFolder, "db.bak")))
+            if (File.Exists(Path.Combine(pDataFolder, "db.bak")))
             {
                 try
                 {
